requirements.txt
openai
gradio
python-dotenv
sqlite-utils   # optional, but helpful if you want extra SQLite tools
